requirements.txt
pydantic
python-dotenv
pytest
pytest-mock
